ENH: Button in notebook gutter to create new cell below current one.
