+++ packages/frontend/src/notebook/notebook_editor.css
@@ -31,7 +31,7 @@
 
 .cell {
     position: relative;
-    margin: 0.7em 0;
+    margin: 0.25em 0;
     display: flex;
     flex-direction: row;
     align-items: baseline;
